aptos_python/pyproject.toml
[tool.poetry]
name = "aptos-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "aptos_python"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
flake8 = "^5.0.4"
aptos-sdk = "^0.4.1"
python-dotenv = "^0.21.0"
python-binance = "^1.0.16"
pyyaml = "^6.0"
pandas = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
